修改菜单

+++ src/common/requestNode.js
@@ -13,23 +13,26 @@ service.interceptors.request.use(config => {
 	config.headers['Authorization'] = localStorage.getItem('token')
 	return config
 }, error => {
-	// Do something with request error
-	console.log(error) // for debug
 	Promise.reject(error)
 })
 
 // respone interceptor
 service.interceptors.response.use(
 	response => {
-		if (   response.data.code == 1001 
-			|| response.data.code == 1002 
-			|| response.data.code == 1003 
-			|| response.data.code == 1004) {
-			localStorage.clear()
-			window.location.href = href()
+		if (response.data.code != 0) {
+			if (   response.data.code == 1001 
+				|| response.data.code == 1002 
+				|| response.data.code == 1003 
+				|| response.data.code == 1004) {
+				localStorage.clear()
+				window.location.href = href()
+				return Promise.reject('error')
+			}
+			Message.error(response.data.msg)
 			return Promise.reject('error')
+		} else {
+			return response
 		}
-		return response
 	},
 	error => {
 		Message.error(error.toString())
